fix: rename backends module to modelling
