--- dspy/backends/__init__.py
@@ -1,3 +1,0 @@
-from .template import TemplateBackend
-from .json import JSONBackend
-from .lm import LiteLM
