+++ tests/evaluate/test_evaluate.py
@@ -1,5 +1,5 @@
 import dspy
-from dspy.backends.template import TemplateBackend
+from dspy.modelling import TemplateBackend
 from dspy.evaluate.evaluate import Evaluate
 from dspy.evaluate.metrics import answer_exact_match
 from dspy.predict import Predict
